# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "RetailConnect",
                 "version":  "1.0"
             },
    "paths":  {
                  "/api/campaigns":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "Campaign"
                                                               ],
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/CreateCampaignRequest"
                                                                                                                              }
                                                                                                               },
                                                                                          "text/json":  {
                                                                                                            "schema":  {
                                                                                                                           "$ref":  "#/components/schemas/CreateCampaignRequest"
                                                                                                                       }
                                                                                                        },
                                                                                          "application/*+json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/CreateCampaignRequest"
                                                                                                                                }
                                                                                                                 }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "201":  {
                                                                                    "description":  "Created",
                                                                                    "content":  {
                                                                                                    "text/plain":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/CampaignResponse"
                                                                                                                                  }
                                                                                                                   },
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/CampaignResponse"
                                                                                                                                        }
                                                                                                                         },
                                                                                                    "text/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/CampaignResponse"
                                                                                                                                 }
                                                                                                                  }
                                                                                                }
                                                                                },
                                                                        "400":  {
                                                                                    "description":  "Bad Request",
                                                                                    "content":  {
                                                                                                    "text/plain":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                  }
                                                                                                                   },
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                        }
                                                                                                                         },
                                                                                                    "text/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                 }
                                                                                                                  }
                                                                                                }
                                                                                }
                                                                    }
                                                  },
                                         "get":  {
                                                     "tags":  [
                                                                  "Campaign"
                                                              ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Success",
                                                                                   "content":  {
                                                                                                   "text/plain":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "array",
                                                                                                                                     "items":  {
                                                                                                                                                   "$ref":  "#/components/schemas/CampaignListItem"
                                                                                                                                               }
                                                                                                                                 }
                                                                                                                  },
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "type":  "array",
                                                                                                                                           "items":  {
                                                                                                                                                         "$ref":  "#/components/schemas/CampaignListItem"
                                                                                                                                                     }
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "array",
                                                                                                                                    "items":  {
                                                                                                                                                  "$ref":  "#/components/schemas/CampaignListItem"
                                                                                                                                              }
                                                                                                                                }
                                                                                                                 }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                     },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CampaignListItem":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "id":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int32"
                                                                                          },
                                                                                   "name":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                   "isActive":  {
                                                                                                    "type":  "boolean"
                                                                                                }
                                                                               },
                                                                "additionalProperties":  false
                                                            },
                                       "CampaignResponse":  {
                                                                "type":  "object",
                                                                "properties":  {
                                                                                   "id":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int32"
                                                                                          },
                                                                                   "name":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                   "startDate":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "date-time",
                                                                                                     "nullable":  true
                                                                                                 },
                                                                                   "endDate":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "date-time",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                   "isActive":  {
                                                                                                    "type":  "boolean"
                                                                                                },
                                                                                   "createdDate":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "date-time"
                                                                                                   }
                                                                               },
                                                                "additionalProperties":  false
                                                            },
                                       "CreateCampaignRequest":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "name":  {
                                                                                                     "type":  "string",
                                                                                                     "nullable":  true
                                                                                                 },
                                                                                        "startDate":  {
                                                                                                          "type":  "string",
                                                                                                          "format":  "date-time",
                                                                                                          "nullable":  true
                                                                                                      },
                                                                                        "endDate":  {
                                                                                                        "type":  "string",
                                                                                                        "format":  "date-time",
                                                                                                        "nullable":  true
                                                                                                    },
                                                                                        "isActive":  {
                                                                                                         "type":  "boolean"
                                                                                                     }
                                                                                    },
                                                                     "additionalProperties":  false
                                                                 },
...
                                       "ProblemDetails":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "type":  {
                                                                                              "type":  "string",
                                                                                              "nullable":  true
                                                                                          },
                                                                                 "title":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                 "status":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int32",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "detail":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "instance":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              }
                                                                             },
                                                              "additionalProperties":  {

                                                                                       }
                                                          },
...
                                   }
                   }
}